+++ basculehttp/enforcer.go
@@ -18,15 +18,14 @@
 package basculehttp
 
 import (
-	"context"
 	"errors"
 	"net/http"
 
-	"github.com/go-kit/kit/log"
-	"github.com/go-kit/kit/log/level"
 	"github.com/justinas/alice"
 	"github.com/xmidt-org/bascule"
+	"github.com/xmidt-org/sallust"
 	"go.uber.org/fx"
+	"go.uber.org/zap"
 )
 
 //go:generate stringer -type=NotFoundBehavior
@@ -54,7 +53,7 @@ type EOptionsIn struct {
 type enforcer struct {
 	notFoundBehavior NotFoundBehavior
 	rules            map[bascule.Authorization]bascule.Validator
-	getLogger        func(context.Context) log.Logger
+	getLogger        sallust.GetLoggerFunc
 	onErrorResponse  OnErrorResponse
 }
 
@@ -63,12 +62,12 @@ func (e *enforcer) decorate(next http.Handler) http.Handler {
 		ctx := request.Context()
 		logger := e.getLogger(ctx)
 		if logger == nil {
-			logger = defaultGetLoggerFunc(ctx)
+			logger = sallust.Get(ctx)
 		}
 		auth, ok := bascule.FromContext(ctx)
 		if !ok {
 			err := errors.New("no authentication found")
-			logger.Log(level.Key(), level.ErrorValue(), errorKey, err.Error())
+			logger.Error(err.Error())
 			e.onErrorResponse(MissingAuthentication, err)
 			response.WriteHeader(http.StatusForbidden)
 			return
@@ -76,9 +75,8 @@ func (e *enforcer) decorate(next http.Handler) http.Handler {
 		rules, ok := e.rules[auth.Authorization]
 		if !ok {
 			err := errors.New("no rules found for authorization")
-			logger.Log(level.Key(), level.ErrorValue(),
-				errorKey, err.Error(), "rules", rules,
-				"authorization", auth.Authorization, "behavior", e.notFoundBehavior)
+			logger.Error(err.Error(), zap.Any("rules", rules),
+				zap.String("authorization", string(auth.Authorization)), zap.Int("behavior", int(e.notFoundBehavior)))
 			switch e.notFoundBehavior {
 			case Forbid:
 				e.onErrorResponse(ChecksNotFound, err)
@@ -94,13 +92,13 @@ func (e *enforcer) decorate(next http.Handler) http.Handler {
 		} else {
 			err := rules.Check(ctx, auth.Token)
 			if err != nil {
-				logger.Log(level.Key(), level.ErrorValue(), errorKey, err)
+				logger.Error(err.Error())
 				e.onErrorResponse(ChecksFailed, err)
 				WriteResponse(response, http.StatusForbidden, err)
 				return
 			}
 		}
-		logger.Log(level.Key(), level.DebugValue(), "msg", "authentication accepted by enforcer")
+		logger.Debug("authentication accepted by enforcer")
 		next.ServeHTTP(response, request)
 	})
 }
@@ -111,7 +109,7 @@ func (e *enforcer) decorate(next http.Handler) http.Handler {
 func NewEnforcer(options ...EOption) func(http.Handler) http.Handler {
 	e := &enforcer{
 		rules:           make(map[bascule.Authorization]bascule.Validator),
-		getLogger:       defaultGetLoggerFunc,
+		getLogger:       sallust.Get,
 		onErrorResponse: DefaultOnErrorResponse,
 	}
 
@@ -146,7 +144,7 @@ func WithRules(key bascule.Authorization, v bascule.Validator) EOption {
 
 // WithELogger sets the function to use to get the logger from the context.
 // If no logger is set, nothing is logged.
-func WithELogger(getLogger func(context.Context) log.Logger) EOption {
+func WithELogger(getLogger sallust.GetLoggerFunc) EOption {
 	return func(e *enforcer) {
 		if getLogger != nil {
 			e.getLogger = getLogger